Refactor le composant Detail : mise à jour du chemin d'importation, ajout de la logique de récupération des données et intégration de ngx-charts pour les visualisations
