+++ src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './pages/home/home.component';
 import { NotFoundComponent } from './pages/not-found/not-found.component';
-import { DetailComponent } from './detail/detail.component';
+import { DetailComponent } from './pages/detail/detail.component';
 
 const routes: Routes = [
   {
@@ -17,10 +17,6 @@ const routes: Routes = [
     path: '**', // wildcard
     component: NotFoundComponent,
   },
-  {
-    path: 'details/:name',
-    component: DetailComponent,
-  }
 ];
 
 @NgModule({
