+++ src/app/core/services/olympic.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { catchError, delay, map, tap } from 'rxjs/operators';
+import { catchError, map, tap } from 'rxjs/operators';
 import { Olympic } from '../models/Olympic';
 
 @Injectable({
@@ -30,20 +30,20 @@ export class OlympicService {
     return this._olympics$.asObservable();
   }
 
-  getOlympicByCountry(country: string): Observable<Olympic | undefined> {
+  getOlympicByCountry(country: string): Observable<Olympic | null> {
     return this._olympics$.pipe(
       map((olympics) => {
         if (olympics) {
           return olympics.find(
             (olympic) => olympic.country.toLowerCase() === country
-          ) || undefined;
+          ) || null;
         }
-        return undefined;
+        return null;
       }),
       catchError((error, caught) => { 
         console.error(error);
         return caught;
-      }),
+      })
     );
   }
 }
