--- src/app/detail/detail.component.html
@@ -1,1 +1,0 @@
-<p>detail works!</p>
