+++ src/app/pages/detail/detail.component.html
@@ -1,26 +1,24 @@
 <a
   routerLink="/"
-  class="fixed top-0 left-0 p-2 md:p-6 text-xl md:text-7xl text-teal-600 hover:scale-105 active:scale-95 cursor-pointer transition-all"
+  class="fixed top-0 left-0 p-6 text-7xl text-teal-600 hover:scale-105 active:scale-95 cursor-pointer transition-all"
   >←</a
 >
-<main class="grid gap-8 w-fit p-2">
+<main class="grid gap-8 w-fit p-2 lg:w-[32rem]">
   <h2
     class="px-6 py-2 bg-teal-700 text-white text-xl lg:text-5xl font-extralight rounded-xl text-center"
   >
-    {{ (olympic$ | async)?.country }}
+    {{ olympicData?.country }}
   </h2>
-  <div class="grid md:flex gap-6 items-center w-full">
+  <div class="flex gap-6 items-center w-full">
     <div
       class="h-full grid gap-2 lg:gap-4 place-content-center rounded-xl border-2 border-teal-700 p-4 w-full"
     >
       <h3 class="md:text-xl lg:text-2xl font-extralight text-center">
         Number of entries
       </h3>
-      <ng-container *ngIf="olympic$ | async; else loading">
-        <span class="text-center font-semibold text-xl lg:text-3xl">{{
-          (olympic$ | async)?.participations?.length
-        }}</span>
-      </ng-container>
+      <span class="text-center font-semibold text-xl lg:text-3xl">{{
+        olympicData?.participations?.length
+      }}</span>
     </div>
     <div
       class="h-full grid gap-4 rounded-xl border-2 border-teal-700 p-4 w-full place-content-center"
@@ -28,11 +26,9 @@
       <h3 class="lg:text-2xl md:text-lg font-extralight text-center">
         Total number medals
       </h3>
-      <ng-container *ngIf="totalMedals$ | async; else loading">
-        <span class="text-center font-semibold text-xl lg:text-3xl">
-          {{ totalMedals$ | async }}
-        </span>
-      </ng-container>
+      <span class="text-center font-semibold text-xl lg:text-3xl">
+        {{ getTotalMedals() }}
+      </span>
     </div>
     <div
       class="h-full grid gap-4 rounded-xl border-2 border-teal-700 p-4 w-full place-content-center"
@@ -40,36 +36,21 @@
       <h3 class="lg:text-2xl md:text-lg font-extralight text-center">
         Total number of athletes
       </h3>
-      <ng-container *ngIf="totalAthletes$ | async; else loading">
-        <span class="text-center font-semibold text-xl lg:text-3xl">
-          {{ totalAthletes$ | async }}
-        </span>
-      </ng-container>
+      <span class="text-center font-semibold text-xl lg:text-3xl">
+        {{ getTotalNumberOfAthletes() }}
+      </span>
     </div>
   </div>
-  <ng-container *ngIf="lineChartData$ | async; else loading">
-    <ngx-charts-line-chart
-      [results]="lineChartData$ | async"
-      [view]="view"
-      [showXAxisLabel]="true"
-      [showYAxisLabel]="true"
-      [xAxis]="true"
-      [yAxis]="true"
-      [autoScale]="true"
-      [animations]="false"
-    >
-    </ngx-charts-line-chart>
-  </ng-container>
-
-  <ng-template #loading>
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 -960 960 960"
-      class="aspect-square h-6 lg:h-10 text-center fill-teal-600 animate-spin mx-auto"
-    >
-      <path
-        d="M480-80q-82 0-155-31.5t-127.5-86Q143-252 111.5-325T80-480q0-83 31.5-155.5t86-127Q252-817 325-848.5T480-880q17 0 28.5 11.5T520-840q0 17-11.5 28.5T480-800q-133 0-226.5 93.5T160-480q0 133 93.5 226.5T480-160q133 0 226.5-93.5T800-480q0-17 11.5-28.5T840-520q17 0 28.5 11.5T880-480q0 82-31.5 155t-86 127.5q-54.5 54.5-127 86T480-80Z"
-      />
-    </svg>
-  </ng-template>
+  @if (olympicData) {
+  <ngx-charts-line-chart
+    [results]="convertOlympicDataToLineChartData(olympicData)"
+    [view]="[500, 300]"
+    [showXAxisLabel]="true"
+    [showYAxisLabel]="true"
+    [xAxis]="true"
+    [yAxis]="true"
+    class="border rounded-xl flex items-center justify-center"
+  >
+  </ngx-charts-line-chart>
+  }
 </main>
